Update afishaList.ftl to complete your code

+++ SpringBootAfisha/src/main/java/ru/lanit/springboot/afisha/repos/AfishaRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface AfishaRepository extends CrudRepository<Afisha, Long> {
+public interface AfishaRepository extends CrudRepository<Afisha, Integer> {
     List<Afisha> findByName(String name);
 }
