+++ SpringBootAfisha/src/main/java/ru/lanit/springboot/afisha/repos/UserRepository.java
@@ -7,6 +7,6 @@ import ru.lanit.springboot.afisha.entities.User;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface UserRepository extends JpaRepository<User, Long>{
+public interface UserRepository extends JpaRepository<User, Integer>{
     User findByUsername(String username);
 }
